requirements.txt
fastapi
uvicorn
streamlit
requests
openai
psycopg[binary]
sqlalchemy
pgvector
pydantic
python-dotenv
sqlglot
tenacity
pandas
numpy